Subir archivos y descargar comprobante

+++ controllers/admin.js
@@ -1,6 +1,9 @@
-const mongoose = require('mongoose');
+// const mongoose = require('mongoose');
 const { validationResult } = require('express-validator');
+const file = require('../utils/file')
 const Producto = require('../models/producto');
+
+const ITEMS_POR_PAGINA = 5;
 
 exports.getCrearProducto = (req, res, next) => {
     res.render('admin/editar-producto', { 
@@ -8,6 +11,7 @@ exports.getCrearProducto = (req, res, next) => {
       path: '/admin/crear-producto',
       modoEdicion: false,
       tieneError: false,
+      autenticado: req.session.autenticado,
       mensajeError: null,
       erroresValidacion: []
     });
@@ -15,9 +19,26 @@ exports.getCrearProducto = (req, res, next) => {
 
 exports.postCrearProducto = (req, res, next) => {
     const nombre = req.body.nombre;
-    const urlImagen = req.body.urlImagen;
+    // const urlImagen = req.body.urlImagen;
+    const imagen = req.file;
     const precio = req.body.precio;
     const descripcion = req.body.descripcion;
+
+    if (!imagen) {
+        return res.status(422).render('admin/editar-producto', {
+            path: '/admin/editar-producto',
+            titulo: 'Crear Producto',
+            modoEdicion: false,
+            tieneError: true,
+            mensajeError: 'No hay imagen de Producto',
+            erroresValidacion: [],
+            producto: {
+                nombre: nombre,
+                precio: precio,
+                descripcion: descripcion
+            },
+        });
+    }
     
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -31,36 +52,27 @@ exports.postCrearProducto = (req, res, next) => {
         erroresValidacion: errors.array(),
         producto: {
           nombre: nombre,
-          urlImagen: urlImagen,
           precio: precio,
           descripcion: descripcion
         }
       });
     }
 
+    const urlImagen = imagen.path;
+
     const producto = new Producto({
-        _id: new mongoose.Types.ObjectId('67202c4ea32f034a44727cfa'),
+        // _id: new mongoose.Types.ObjectId('67202c4ea32f034a44727cfa'),
         nombre: nombre,
         precio: precio,
         descripcion: descripcion,
         urlImagen: urlImagen,
         idUsuario: req.usuario._id
     });
-    // const producto = new Producto({
-    //     nombre: nombre,
-    //     precio: precio,
-    //     descripcion: descripcion,
-    //     urlImagen: urlImagen,
-    //     idUsuario: req.usuario._id
-    // });
     producto.save()
         .then(result => {
-            console.log(result);
             res.redirect('/admin/productos');
         })
         .catch(err => {
-            // console.log(err);
-            // res.redirect('/500');
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -84,6 +96,7 @@ exports.getEditarProducto = (req, res, next) => {
                 titulo: 'Editar Producto',
                 path: '/admin/edit-producto',
                 modoEdicion: modoEdicion,
+                autenticado: req.session.autenticado,
                 producto: producto,
                 tieneError: false,
                 mensajeError: null,
@@ -102,7 +115,8 @@ exports.postEditarProducto = (req, res, next) => {
     const idProducto = req.body.idProducto;
     const nombre = req.body.nombre;
     const precio = req.body.precio;
-    const urlImagen = req.body.urlImagen;
+    // const urlImagen = req.body.urlImagen;
+    const imagen = req.file;
     const descripcion = req.body.descripcion;
 
     const errors = validationResult(req);
@@ -132,7 +146,10 @@ exports.postEditarProducto = (req, res, next) => {
             producto.nombre = nombre;
             producto.precio = precio;
             producto.descripcion = descripcion;
-            producto.urlImagen = urlImagen;
+            if (imagen) {
+                file.deleteFile(producto.urlImagen);
+                producto.urlImagen = imagen.path;
+            }
             return producto.save();
         })
         .then(result => {
@@ -149,15 +166,30 @@ exports.postEditarProducto = (req, res, next) => {
 
 
 exports.getProductos = (req, res, next) => {
+    const pagina = +req.query.pagina || 1;
+    let itemsTotales;
     Producto
         // .find()
         .find({ idUsuario: req.usuario._id })
+        .countDocuments()
+        .then(numProductos => {
+        itemsTotales = numProductos;
+        return Producto.find()
+            .skip((pagina - 1) * ITEMS_POR_PAGINA)
+            .limit(ITEMS_POR_PAGINA);
+        })
         .then(productos => {
             res.render('admin/productos', {
                 prods: productos,
                 titulo: "Administracion de Productos",
                 path: '/admin/productos',
-                autenticado: req.session.autenticado
+                autenticado: req.session.autenticado,
+                paginaActual: pagina,
+                tienePaginaSiguiente: ITEMS_POR_PAGINA * pagina < itemsTotales,
+                tienePaginaAnterior: pagina > 1,
+                paginaSiguiente: pagina + 1,
+                paginaAnterior: pagina - 1,
+                ultimaPagina: Math.ceil(itemsTotales / ITEMS_POR_PAGINA)
             });
         })
         .catch(err => {
@@ -170,7 +202,15 @@ exports.getProductos = (req, res, next) => {
 
 exports.postEliminarProducto = (req, res, next) => {
     const idProducto = req.body.idProducto;
-    Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id }) // es una condición donde verifica si los id mandados son iguales
+    // Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id }) // es una condición donde verifica si los id mandados son iguales
+    Producto.findById(idProducto)
+        .then(producto => {
+            if(!producto) {
+                return next(new Error('Producto no se ha encontrado'));
+            }
+            file.deleteFile(producto.urlImagen);
+            return Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id });
+        })
         .then(result => {
             console.log('Producto eliminado satisfactoriamente');
             res.redirect('/admin/productos');
@@ -181,3 +221,23 @@ exports.postEliminarProducto = (req, res, next) => {
             return next(error);
         });
 }; 
+
+
+exports.deleteProducto = (req, res, next) => {
+    const idProducto = req.params.idProducto;
+    Producto.findById(idProducto)
+        .then(producto => {
+            if (!producto) {
+            return next(new Error('Producto no encontrado'));
+            }
+            file.deleteFile(producto.urlImagen);
+            return Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id });
+        })
+        .then(() => {
+            console.log('PRODUCTO ELIMINADO');
+            res.status(200).json({ message: 'Exitoso' });
+        })
+        .catch(err => {
+            res.status(500).json({ message: 'Eliminacion del producto fallo' });
+        });
+};
